+++ src/basic/services/orderService.js
@@ -1,6 +1,6 @@
 import { orderStore } from "../store/orderStore.js";
-import { POINTS, POINTS_QUANTITY_THRESHOLDS, QUANTITY_BONUS_POINTS } from "../constants/index.js";
 import { discountService } from "./discountService.js";
+import { pointService } from "./pointService.js";
 
 export class OrderService {
   constructor() {
@@ -20,7 +20,7 @@ export class OrderService {
    * 포인트를 계산하고 업데이트합니다.
    */
   calculatePoints(cartItems, totalAmount, isTuesday, itemCount) {
-    const pointsData = this.calculateLoyaltyPoints(cartItems, totalAmount, isTuesday, itemCount);
+    const pointsData = pointService.calculatePoints(cartItems, totalAmount, isTuesday, itemCount);
     this.orderStore.updatePoints(pointsData);
     return pointsData;
   }
@@ -48,59 +48,6 @@ export class OrderService {
       })),
       isTuesday,
     };
-  }
-
-  /**
-   * 포인트를 계산합니다.
-   */
-  calculateLoyaltyPoints(cartItems, totalAmount, isTuesday, itemCount) {
-    if (cartItems.length === 0) {
-      return { totalPoints: 0, details: [] };
-    }
-
-    const basePoints = Math.floor(totalAmount / POINTS.BASE_RATE);
-    let finalPoints = 0;
-    const pointsDetail = [];
-
-    if (basePoints > 0) {
-      finalPoints = basePoints;
-      pointsDetail.push(`기본: ${basePoints}p`);
-    }
-
-    // 화요일 2배
-    if (isTuesday && basePoints > 0) {
-      finalPoints = basePoints * POINTS.TUESDAY_MULTIPLIER;
-      pointsDetail.push("화요일 2배");
-    }
-
-    // 세트 구매 보너스
-    const hasKeyboard = cartItems.some(item => item.id === "p1");
-    const hasMouse = cartItems.some(item => item.id === "p2");
-    const hasMonitorArm = cartItems.some(item => item.id === "p3");
-
-    if (hasKeyboard && hasMouse) {
-      finalPoints += POINTS.KEYBOARD_MOUSE_SET;
-      pointsDetail.push("키보드+마우스 세트 +50p");
-    }
-
-    if (hasKeyboard && hasMouse && hasMonitorArm) {
-      finalPoints += POINTS.FULL_SET;
-      pointsDetail.push("풀세트 구매 +100p");
-    }
-
-    // 수량 보너스
-    if (itemCount >= POINTS_QUANTITY_THRESHOLDS.LARGE_BULK) {
-      finalPoints += QUANTITY_BONUS_POINTS.LARGE_BULK;
-      pointsDetail.push("대량구매(30개+) +100p");
-    } else if (itemCount >= POINTS_QUANTITY_THRESHOLDS.MEDIUM_BULK) {
-      finalPoints += QUANTITY_BONUS_POINTS.MEDIUM_BULK;
-      pointsDetail.push("대량구매(20개+) +50p");
-    } else if (itemCount >= POINTS_QUANTITY_THRESHOLDS.SMALL_BULK) {
-      finalPoints += QUANTITY_BONUS_POINTS.SMALL_BULK;
-      pointsDetail.push("대량구매(10개+) +20p");
-    }
-
-    return { totalPoints: finalPoints, details: pointsDetail };
   }
 
   /**
